Setup a tinc configuration for the home server.
tinc.conf configures this server to connect to the cloud server and defines the network interface to use for tinc.
hosts/home contains the public key and IP address of this system.
tinc-up and tinc-down are scripts which bind tinc to the network interface.
rsa_key.priv is the private key.